Fixes error when asset_host is not a string.
Rails' asset_host can be either a string or an object that respond to
`call` (e.g. Proc). There was an error because when it was a Proc it
was trying to concatenate it as a string.

Since at assets:precompile we don't know about the request to pass it
forward to the Proc/Class, the better we can do is probably use the
relative path.