Offer an install strategy that does not preserve modes
When compiling assets as different users (on a typical Linux file
system), the `preserve` flag used in the Copy strategy can throw
permission errors as it tries to preserve file mode/timestamp
information. This `copy_no_preserve` strategy offers an alternative. It
may be better to simply remove the flag, but this approach offers a choice.