Check content css file against regex to avoid calling through ActionController for the stylesheet path
TinyMCE allows for [relative urls for content_css](https://www.tinymce.com/docs/configure/content-appearance/#content_css), however the transformer first checks with `ActionController::Base` to find the path for the given file, which can raise an error.

This adds a check in front in case the file given is a relative uri, and passes that through to the `tinymce` configuration before going through the rails/sprockets pipeline.