Only include vendored assets in the gem
With the introduction of the github workflow in
a432faf22e86577e69949edb39a37a52763ec458 bundler is run as part of
making the gem and this adds bundled code in vendor/bundle.

With the "vendor/**/*" spec in the gemspec file this code also gets
included in the gem by accident.