Make sure this library works with the new Propshaft manifest format (#320)
* Make sure this library works with the new Propshaft manifest format

Since Propshaft 1.2.0, the manifest format has changed to define a
list of properties for each logical path. The format of the manifest
was never public, but I'm aware that some code were parsing that file
and trying to use it. That is why `Propshaft::Manifest` class was added.

This change uses that class instead of parsing the manifest file.